+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 New Transformers:
 
-* r/type/lowercase_string
-* r/type/uppercase_string
+* r/format/lowercase
+* r/format/uppercase
 
 # 1.3.0 (November 4th, 2020)